# docs/swagger.yaml
definitions:
...
  github_com_maidulcu_masaar-crm_internal_domain.InvoiceStatus:
    enum:
    - draft
    - sent
    - paid
    type: string
    x-enum-varnames:
    - InvoiceDraft
    - InvoiceSent
    - InvoicePaid
...
  github_com_maidulcu_masaar-crm_internal_domain.VATInvoice:
    properties:
      deal_id:
        type: string
      id:
        type: string
      invoice_no:
        type: string
      issued_at:
        type: string
      qr_payload:
        type: string
      status:
        $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.InvoiceStatus'
      subtotal:
        type: number
      total:
        type: number
      vat_amount:
        type: number
      vat_rate:
        type: number
    type: object
...
info:
  contact: {}
paths:
...
  /invoices:
    post:
      consumes:
      - application/json
      description: Generates a VAT invoice (5% UAE VAT) for a deal. Invoice number
        is auto-assigned as INV-YYYY-NNNN.
      parameters:
      - description: Invoice payload
        in: body
        name: body
        required: true
        schema:
          properties:
            deal_id:
              type: string
            subtotal:
              type: number
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.VATInvoice'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create VAT invoice
      tags:
      - Invoices
  /invoices/{id}:
    get:
      parameters:
      - description: Invoice UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.VATInvoice'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get invoice
      tags:
      - Invoices
  /invoices/{id}/send:
    post:
      description: Transitions the invoice status from draft → sent.
      parameters:
      - description: Invoice UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark invoice as sent
      tags:
      - Invoices
  /invoices/{id}/status:
    patch:
      consumes:
      - application/json
      description: Manually set invoice status to draft|sent|paid.
      parameters:
      - description: Invoice UUID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: body
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update invoice status
      tags:
      - Invoices
...
swagger: "2.0"